--- a/compiler/AST/build.cpp
+++ b/compiler/AST/build.cpp
@@ -1529,6 +1529,11 @@ void addTaskIntent(CallExpr* ti, Expr* var, IntentTag intent, Expr* ri) {
   }
 }
 
+static CallExpr* copyByrefVars(CallExpr* byrefVarsSource) {
+  if (!byrefVarsSource) return NULL;
+  return byrefVarsSource->copy();
+}
+
 static void
 addByrefVars(BlockStmt* target, CallExpr* byrefVarsSource) {
   // nothing to do if there is no 'ref' clause
@@ -1547,6 +1552,102 @@ addByrefVars(BlockStmt* target, CallExpr* byrefVarsSource) {
   // will be automatically resolved in resolve().
 }
 
+
+// Build up a "lowered" coforall loop. We lower coforalls into for-loops with
+// explicit fork-join task creation via an EndCount.
+static BlockStmt* buildLoweredCoforall(Expr* indices,
+                                       VarSymbol* iterator,
+                                       CallExpr* byref_vars,
+                                       BlockStmt* body,
+                                       bool zippered,
+                                       bool bounded) {
+
+  BlockStmt* taskBlk = new BlockStmt();
+  taskBlk->insertAtHead(body);
+
+  VarSymbol* coforallCount = newTempConst("_coforallCount");
+  VarSymbol* numTasks = newTemp("numTasks");
+  VarSymbol* useLocalEndCount = gTrue;
+  VarSymbol* countRunningTasks = gTrue;
+
+  BlockStmt* onBlock = findStmtWithTag(PRIM_BLOCK_ON, body);
+  // For remote coforalls (e..g. coforall indices in iterator do on indices) we
+  // just do a remote fork instead of creating a task locally. Do not count
+  // running tasks locally, and use network atomic EndCounts if available
+  if (onBlock) {
+    onBlock->blockInfoGet()->primitive = primitives[PRIM_BLOCK_COFORALL_ON];
+    onBlock->insertAtTail(new CallExpr("_downEndCount", coforallCount));
+    addByrefVars(onBlock, byref_vars);
+    taskBlk->blockTag = BLOCK_SCOPELESS;
+    useLocalEndCount = gFalse;
+    countRunningTasks = gFalse;
+  } else {
+    taskBlk->blockInfoSet(new CallExpr(PRIM_BLOCK_COFORALL));
+    taskBlk->insertAtTail(new CallExpr("_downEndCount", coforallCount));
+    addByrefVars(taskBlk, byref_vars);
+  }
+
+  BlockStmt* block = ForLoop::buildForLoop(indices, new SymExpr(iterator), taskBlk, true, zippered);
+  if (bounded) {
+    block->insertAtHead(new CallExpr("_upEndCount", coforallCount, countRunningTasks, numTasks));
+    block->insertAtHead(new CallExpr(PRIM_MOVE, numTasks, new CallExpr(".", iterator,  new_CStringSymbol("size"))));
+    block->insertAtHead(new DefExpr(numTasks));
+    block->insertAtTail(new CallExpr("_waitEndCount", coforallCount, countRunningTasks, numTasks));
+  } else {
+    taskBlk->insertBefore(new CallExpr("_upEndCount", coforallCount, countRunningTasks));
+    block->insertAtTail(new CallExpr("_waitEndCount", coforallCount, countRunningTasks));
+  }
+  block->insertAtTail(new CallExpr("_endCountFree", coforallCount));
+
+  block->insertAtHead(new CallExpr(PRIM_MOVE, coforallCount, new CallExpr("_endCountAlloc", useLocalEndCount)));
+  block->insertAtHead(new DefExpr(coforallCount));
+  return block;
+}
+
+
+// Build up AST for coforalls. For something like:
+//
+//     coforall indices in iterator with (byref_vars) { body(); }
+//
+// This effectively builds up:
+//
+//     var tmpIter = iterator;
+//     param bounded = isBoundedRange(tmpIter) || isDomain(tmpIter) || isArray(tmpIter);
+//     param useLocalEndCount, countRunningTasks = !bodyContainsOnStmt();
+//     if bounded {
+//       var numTasks = tmpIter.size;
+//       var _coforallCount = _endCountAlloc(useLocalEndCount);
+//       // only bump EndCount once, instead of once per task
+//       _upEndCount(_coforallCount, countRunningTasks, numTasks);
+//       for indices in tmpIter {
+//         /* PRIM_BLOCK_COFORALL (byref_vars) */ {
+//           body();
+//           _downEndCount(_coforallCount);
+//         }
+//       }
+//       _waitEndCount(_coforallCount, countRunningTasks, numTasks);
+//       _endCountFree(_coforallCount);
+//     } else {
+//       var _coforallCount = _endCountAlloc(useLocalEndCount);
+//       for indices in tmpIter {
+//         _upEndCount(_coforallCount, countRunningTasks);
+//         /* PRIM_BLOCK_COFORALL (byref_vars) */ {
+//           body();
+//           _downEndCount(_coforallCount);
+//         }
+//       }
+//       _waitEndCount(_coforallCount, countRunningTasks);
+//       _endCountFree(_coforallCount);
+//     }
+//
+// For coforall+ons:
+//
+//     coforall indices in iterator do on indices{ body(); }
+//
+// there are some minor differences. We use network atomics for the EndCount if
+// they're available, we won't manipulate here.runningTaskCount, and we'll use
+// PRIM_BLOCK_COFORALL_ON instead of PRIM_BLOCK_COFORALL so that we just do
+// remote-forks instead of creating any tasks locally.
 BlockStmt* buildCoforallLoopStmt(Expr* indices,
                                  Expr* iterator,
                                  CallExpr* byref_vars,
@@ -1555,108 +1656,38 @@ BlockStmt* buildCoforallLoopStmt(Expr* indices,
 {
   checkControlFlow(body, "coforall statement");
 
-  //
   // insert temporary index when elided by user
-  //
   if (!indices)
     indices = new UnresolvedSymExpr("chpl__elidedIdx");
-
   checkIndices(indices);
-
-  //
-  // detect on-statement directly inside coforall-loop
-  //
-  BlockStmt* onBlock = findStmtWithTag(PRIM_BLOCK_ON, body);
-
+  
   SET_LINENO(body);
 
-  if (onBlock) {
-    //
-    // optimization of on-statements directly inside coforall-loops
-    //
-    //   In this case, the on-statement is made into a non-blocking
-    //   on-statement and the coforall is serialized (rather than
-    //   wasting threads that would do nothing other than wait on the
-    //   on-statement.
-    //
-    VarSymbol* coforallCount = newTempConst("_coforallCount");
-    BlockStmt* block = ForLoop::buildForLoop(indices, iterator, body, true, zippered);
-    block->insertAtHead(new CallExpr(PRIM_MOVE, coforallCount, new CallExpr("_endCountAlloc", /* forceLocalTypes= */gFalse)));
-    block->insertAtHead(new DefExpr(coforallCount));
-    body->insertAtHead(new CallExpr("_upEndCount", coforallCount, gFalse));
-    block->insertAtTail(new CallExpr("_waitEndCount", coforallCount, gFalse));
-    block->insertAtTail(new CallExpr("_endCountFree", coforallCount));
-    onBlock->blockInfoGet()->primitive = primitives[PRIM_BLOCK_COFORALL_ON];
-    addByrefVars(onBlock, byref_vars);
-    BlockStmt* innerOnBlock = new BlockStmt();
-    for_alist(tmp, onBlock->body) {
-      innerOnBlock->insertAtTail(tmp->remove());
-    }
-    onBlock->insertAtHead(innerOnBlock);
-    onBlock->insertAtTail(new CallExpr("_downEndCount", coforallCount));
-    return block;
-  } else {
+  VarSymbol* tmpIter = newTemp("tmpIter");
+  tmpIter->addFlag(FLAG_EXPR_TEMP);
+  tmpIter->addFlag(FLAG_MAYBE_REF);
 
-    BlockStmt* coforallBlk = new BlockStmt();
+  BlockStmt* coforallBlk = new BlockStmt();
+  coforallBlk->insertAtTail(new DefExpr(tmpIter));
+  coforallBlk->insertAtTail(new CallExpr(PRIM_MOVE, tmpIter, iterator));
 
-    BlockStmt* vectorCoforallBlk = new BlockStmt();
-    BlockStmt* nonVectorCoforallBlk = new BlockStmt();
+  BlockStmt* vectorCoforallBlk = buildLoweredCoforall(indices, tmpIter, copyByrefVars(byref_vars), body->copy(), zippered, /*bounded=*/true);
+  BlockStmt* nonVectorCoforallBlk = buildLoweredCoforall(indices, tmpIter, byref_vars, body, zippered, /*bounded=*/false);
 
-    VarSymbol* tmpIter = newTemp("tmpIter");
-    tmpIter->addFlag(FLAG_EXPR_TEMP);
-    tmpIter->addFlag(FLAG_MAYBE_REF);
-    coforallBlk->insertAtTail(new DefExpr(tmpIter));
-    coforallBlk->insertAtTail(new CallExpr(PRIM_MOVE, tmpIter, iterator));
-    {
-      VarSymbol* coforallCount = newTempConst("_coforallCount");
-      BlockStmt* beginBlk = new BlockStmt();
-      beginBlk->blockInfoSet(new CallExpr(PRIM_BLOCK_COFORALL));
-      addByrefVars(beginBlk, byref_vars);
-      beginBlk->insertAtHead(body);
-      beginBlk->insertAtTail(new CallExpr("_downEndCount", coforallCount));
-      BlockStmt* block = ForLoop::buildForLoop(indices, new SymExpr(tmpIter), beginBlk, true, zippered);
-      block->insertAtHead(new CallExpr(PRIM_MOVE, coforallCount, new CallExpr("_endCountAlloc", /*forceLocalTypes=*/gTrue)));
-      block->insertAtHead(new DefExpr(coforallCount));
-      beginBlk->insertBefore(new CallExpr("_upEndCount", coforallCount));
-      block->insertAtTail(new CallExpr("_waitEndCount", coforallCount));
-      block->insertAtTail(new CallExpr("_endCountFree", coforallCount));
-      nonVectorCoforallBlk->insertAtTail(block);
-    }
-    {
-      VarSymbol* coforallCount = newTempConst("_coforallCount");
-      BlockStmt* beginBlk = new BlockStmt();
-      beginBlk->blockInfoSet(new CallExpr(PRIM_BLOCK_COFORALL));
-      addByrefVars(beginBlk, byref_vars);
-      beginBlk->insertAtHead(body->copy());
-      beginBlk->insertAtTail(new CallExpr("_downEndCount", coforallCount));
-      VarSymbol* numTasks = newTemp("numTasks");
-      vectorCoforallBlk->insertAtTail(new DefExpr(numTasks));
-      vectorCoforallBlk->insertAtTail(new CallExpr(PRIM_MOVE, numTasks, new CallExpr(".", tmpIter,  new_CStringSymbol("size"))));
-      vectorCoforallBlk->insertAtTail(new CallExpr("_upEndCount", coforallCount, /*countRunningTasks=*/gTrue, numTasks));
-      BlockStmt* block = ForLoop::buildForLoop(indices, new SymExpr(tmpIter), beginBlk, true, zippered);
-      vectorCoforallBlk->insertAtHead(new CallExpr(PRIM_MOVE, coforallCount, new CallExpr("_endCountAlloc", /*forceLocalTypes=*/gTrue)));
-      vectorCoforallBlk->insertAtHead(new DefExpr(coforallCount));
-      block->insertAtTail(new CallExpr("_waitEndCount", coforallCount, /*countRunningTasks=*/gTrue, numTasks));
-      block->insertAtTail(new CallExpr("_endCountFree", coforallCount));
-      vectorCoforallBlk->insertAtTail(block);
-    }
+  VarSymbol* isRngDomArr = newTemp("isRngDomArr");
+  isRngDomArr->addFlag(FLAG_MAYBE_PARAM);
+  coforallBlk->insertAtTail(new DefExpr(isRngDomArr));
 
-    VarSymbol* isRngDomArr  = newTemp("isRngDomArr");
-    isRngDomArr->addFlag(FLAG_MAYBE_PARAM);
-    coforallBlk->insertAtTail(new DefExpr(isRngDomArr));
+  coforallBlk->insertAtTail(new CallExpr(PRIM_MOVE, isRngDomArr,
+                            new CallExpr("||", new CallExpr("isBoundedRange", tmpIter),
+                            new CallExpr("||", new CallExpr("isDomain", tmpIter), new CallExpr("isArray", tmpIter)))));
 
-    coforallBlk->insertAtTail(new CallExpr(PRIM_MOVE, isRngDomArr,
-                              new CallExpr("||", new CallExpr("isBoundedRange", tmpIter),
-                              new CallExpr("||", new CallExpr("isDomain", tmpIter), new CallExpr("isArray", tmpIter)))));
-
-
-    coforallBlk->insertAtTail(new CondStmt(new SymExpr(isRngDomArr),
-                                           vectorCoforallBlk->copy(),
-                                           nonVectorCoforallBlk->copy()));
-
-    return coforallBlk;
-  }
+  coforallBlk->insertAtTail(new CondStmt(new SymExpr(isRngDomArr),
+                                         vectorCoforallBlk,
+                                         nonVectorCoforallBlk));
+  return coforallBlk;
 }
+
 
 BlockStmt* buildParamForLoopStmt(const char* index, Expr* range, BlockStmt* stmts) {
   VarSymbol* indexVar = new VarSymbol(index);
@@ -2869,7 +2900,7 @@ buildCobeginStmt(CallExpr* byref_vars, BlockStmt* block) {
     beginBlk->blockInfoSet(new CallExpr(PRIM_BLOCK_COBEGIN));
     beginBlk->astloc = stmt->astloc;
     // the original byref_vars is dead - will be clean_gvec-ed
-    addByrefVars(beginBlk, byref_vars ? byref_vars->copy() : NULL);
+    addByrefVars(beginBlk, copyByrefVars(byref_vars));
     stmt->insertBefore(beginBlk);
     beginBlk->insertAtHead(stmt->remove());
     beginBlk->insertAtTail(new CallExpr("_downEndCount", cobeginCount));