Optimize remote task creation speed. This applies the same optimization done
for local coforalls in #5018, but to remote coforalls

We used to transform code like:

```chapel
coforall loc in Locales do on loc { }
```

into roughly something like:

```chapel
for loc in Locales {
  _upEndCount(_coforallCount);
  executeOnNB(loc, body());
}
_waitEndCount(_coforallCount);
```

and now we transform it into something like:

```chapel
var tmpIter = Locales;
var bounded = isBoundRange(tmpIter) || isDomain(tmpIter) || isArray(tmpIter);
if bounded {
  var numTasks = tmpIter.size;
  // only bump EndCount once, instead of once per task
  _upEndCount(_coforallCount, numTasks);
  for loc in tmpIter {
    executeOnNB(loc, body());
  }
  _waitEndCount(_coforallCount, numTasks);
} else  {
  // old style coforall+on
}
```

Under ugni this turns 2*numTasks network atomic updates into a single atomic
update. We now only update our `EndCount.i` once, and we don't have to touch
`EndCount.taskCnt` at all.

This pretty dramatically improves remote-task creation, especially under ugni.
For the remote task microbenchmark we see a 2.5X speedup:

|               | before | now   |
| ------------- | ------ | ----- |
| u+q AMO       | 36.5s  | 13.5s |
| u+q processor | 15.5s  | 14.5s |
| gn-aries      | 33.5s  | 32.5s |
| gn-mpi        | 43.5s  | 42.5s |


This results in some decent performance improvements for real benchmarks as
well: 25% speedup for lulesh, up to 50% perf boost for stencil PRK, and minor
improvements for some other benchmarks as well:

| Benchmark   | before       | now          |
| ----------- | ------------ | ------------ |
| lulesh      | 3.35s        | 2.5s         |
| stencil prk | 120 GFlops/s | 130 GFlops/s |
| stencil-opt | 208 GFlops/s | 320 GFlop/s  |
| prom stream | 450 GB/s     | 640 GB/s     |

This resolves #6357