@ronawho 2017-06-08
Passes full local, gasnet, and valgrind testing. Going to do a --verify run as well out of paranoia.

I'll also rebase before merging, but I figured the history might be useful during review in case you want to see what I did to resolve some of the issues I was originally running into